changed datatypes

+++ C libraries/SPI.c
@@ -33,12 +33,12 @@
 extern volatile INT16S pos_var;
 extern volatile INT16S pwm_var;
 
-int16_t control_1_pos;
-int16_t control_1_vel;
-int16_t control_1_cur;
-int16_t control_2_pos;
-int16_t control_2_vel;
-int16_t control_2_cur;
+float control_1_pos;
+float control_1_vel;
+float control_1_cur;
+float control_2_pos;
+float control_2_vel;
+float control_2_cur;
 
 SemaphoreHandle_t POS_1_SEM;
 SemaphoreHandle_t VEL_1_SEM;
