+++ FreeRTOS_position-controller/setup.c
@@ -6,19 +6,19 @@
 
 
 /********************** External declaration of Variables ******************/
-int16_t control_1_pos;
-int16_t control_1_vel;
-int16_t control_1_cur;
-int16_t control_2_pos;
-int16_t control_2_vel;
-int16_t control_2_cur;
+float control_1_pos;
+float control_1_vel;
+float control_1_cur;
+float control_2_pos;
+float control_2_vel;
+float control_2_cur;
 
-int16_t control_1_pos_ref;
-int16_t control_1_vel_ref;
-int16_t control_1_cur_ref;
-int16_t control_2_pos_ref;
-int16_t control_2_vel_ref;
-int16_t control_2_cur_ref;
+float control_1_pos_ref;
+float control_1_vel_ref;
+float control_1_cur_ref;
+float control_2_pos_ref;
+float control_2_vel_ref;
+float control_2_cur_ref;
 
 int16_t glob_protocol;
